Add new module for C++ vector of KTfwd::popgenmut
