+++ fwdpy11/__init__.py.in
@@ -25,6 +25,7 @@ if sys.version_info[0] < 3:
 __version__ = '@PACKAGE_VERSION@'
 from .fwdpp_types import *
 from ._opaque_gametes import *
+from ._opaque_mutations import *
 from .fwdpy11_types import *
 from .regions import *
 from .dev import *
